refactor: usePanel, useTabs 분리
