+++ src/lib/Tabs/hooks/useTabs.tsx
@@ -1,39 +1,12 @@
-import { Children, ReactElement, ReactNode, useEffect, useState } from 'react';
-import { Label, PanelList } from '../../types/type';
+import { useContext } from 'react';
+import { TabsContext } from '../TabsProvider';
 
-const useTabs = (children: ReactNode) => {
-  const [panelList, setPanelList] = useState<PanelList>([]);
-  const [selectedPanelLabel, setSelectedPanelLabel] = useState<Label | null>(
-    null
-  );
+const useTabs = () => {
+  const value = useContext(TabsContext);
 
-  const selectedPanel = panelList.find(
-    ({ label }) => label === selectedPanelLabel
-  );
+  if (!value) throw new Error('tab provider 오류');
 
-  const selectPanel = (label: Label) => setSelectedPanelLabel(label);
-
-  const isSelected = (label: Label) => label === selectedPanelLabel;
-
-  const isDuplicationLabel =
-    new Set(panelList.map((panel) => panel.label)).size !== panelList.length;
-
-  if (isDuplicationLabel) throw new Error('탭의 라벨이 중복되었습니다.');
-
-  useEffect(() => {
-    Children.forEach(children, (child, index) => {
-      const panel = child as ReactElement;
-
-      const { label, disabled } = panel.props;
-
-      setPanelList((prev) => [...prev, { label, disabled, contents: panel }]);
-      if (index === 0) setSelectedPanelLabel(label);
-    });
-
-    return () => setPanelList([]);
-  }, [children]);
-
-  return { panelList, selectedPanel, isSelected, selectPanel };
+  return value;
 };
 
 export default useTabs;
